Extract Scorer read-only preview path from GovernancePipeline
Move the entire read side into a Scorer collaborator: score_tool_call,
score_event (was _score_event), score_tool_call_event, preflight_event,
_meta_to_trace, _persist_score, and the fail-closed _fail_closed fallback.
Scorer previews assessment against a detached clone via Phase1 + the Assessor
and never mutates persisted state; it is injected with context, phase1,
assessor, registry, store, and codec.

The Shield now scores through self._scorer.score_event (bound method) instead of
the pipeline's old _score_event. The pipeline keeps score_tool_call,
score_tool_call_event, and preflight_event as thin forwards. uuid and
RecommendedAction/RiskRecommendation are no longer used in pipeline and were
dropped from its imports.

Two tests are updated deliberately to target the real collaborator seams the
Scorer now owns: test_bridge patches the shared ContextBuilder
(pipeline._context.from_session_event) and test_score_tool_call patches
pipeline._scorer.preflight_event. Three tests now import RecommendedAction from
the canonical tracemill.governance.results. No behavior change: 1884 passed /
4 skipped.
